better strand management for cds2rna_blat

+++ cds2rna_blat.pl
@@ -4,7 +4,8 @@ use Getopt::Std;
 use IPC::Open2;
 
 my $usage = q/Usage:
-  cds2rna_blat.pl tmappings.fa refprot.fa.cidx > tmappings_wCDS.tlf
+  cds2rna_blat.pl [-w outblat.psl] [-b inblat.psl] \
+                  tmappings.fa refprot.fa.cidx > tmappings_wCDS.tlf
   
   tmappings.fa must be a transcripts sequence file 
                  generated with gffread -W -w
@@ -13,13 +14,15 @@ my $usage = q/Usage:
   Output will be a CDS annotated compact gff3 (tlf)
 /;
 umask 0002;
-getopts('Do:') || die($usage."\n");
+getopts('w:b:Do:') || die($usage."\n");
 my $outfile=$Getopt::Std::opt_o;
 if ($outfile) {
   open(OUTF, '>'.$outfile) || die("Error creating output file $outfile\n");
   select(OUTF);
 }
 my $debug=$Getopt::Std::opt_D;
+my $wblat=$Getopt::Std::opt_w;
+my $inblat=$Getopt::Std::opt_b;
 # --
 die($usage."\n") unless @ARGV==2;
 my ($tfa, $pepdb)=@ARGV;
@@ -27,17 +30,26 @@ die("${usage}File $tfa not found!\n") unless -f $tfa;
 die("${usage}File $pepdb not found!\n") unless -f $pepdb;
 pop(@ARGV);
 my ($tid, $defline, $tseq);
-my %tdescr; # tid => description
+my %tdescr; # tid => [prot_len, description]
 my %tm; #a transcript can have multiple mappings (e.g. gmap)
         #we may have multiple blat alignments
         # tid=>[ [tseqid, chr, strand, start, end, [exons], bscore, CDstart, CDend, bstatus], ...]
         #          0       1      2     3      4       5       6       7      8       9
 my @stopCodons=qw(TAA TAG TAR TGA TRA);
 my %stops;
+my $totalprot=0;
 @stops{@stopCodons}=('*') x @stopCodons;
 mkdir('/dev/shm/gpertea');
 my $ftmp="/dev/shm/gpertea/tpep.$$.fa";
 my $counter=0;
+if ($inblat) {
+ die("Error: input blat file must be different from output blat file!\n") if $wblat eq $inblat;
+ open(BLAT, $inblat) || die("Error opening psl file $inblat !\n");
+}
+
+if ($wblat) {
+ open(WBLAT, ">$wblat") || die("Error: cannot create file $wblat !\n");
+}
 {
  local $/="\n>";
  while (<>) {
@@ -83,7 +95,7 @@ my $counter=0;
          $pdescr=$1;
          $pdescr=~s/\.? \(\S+\)$//;
        }
-       $tdescr{$tid}=$pdescr if $pdescr;
+       else { $pdescr=''; }
      }
      else {
        die("Error getting defline from $ftmp\n");
@@ -93,27 +105,38 @@ my $counter=0;
        $pseq.=$_;
      }
      $plen=length($pseq);
+     $totalprot+=$plen; # reference protein space
+     $tdescr{$tid}=[$plen, $pdescr];
      close(PEP);
     } #reading protein
     #--
-    my $blatcmd="blat -t=dnax -q=prot -out=psl -noHead stdin $ftmp stdout";
-    my $prcid=open2(\*BL_OUT, \*DNA_IN, $blatcmd) || die("Error: open2() failed $!\n");
-    print DNA_IN ">$tseqid\n$trawseq\n";
-    close(DNA_IN); #flush input to blat command
+    my $prcid;
+    unless ($inblat) {
+      my $blatcmd="blat -t=dnax -q=prot -out=psl -noHead stdin $ftmp stdout";
+      print STDERR "Running blat cmd:\n$blatcmd\n" if $debug;
+      $prcid=open2(\*BLAT, \*DNA_IN, $blatcmd) || die("Error: open2() failed $!\n");
+      print DNA_IN ">$tseqid\n$trawseq\n";
+      close(DNA_IN); #flush input to blat command
+    }
     #now we can read the output
     {
      local $/="\n";
      # blat PSL format
      # #match #mism #repm #Ns #qgaps #qgapb #tgaps #tgapb strand q_id q_len q_start q_end t_name t_len t_start t_end #blocks blkSizes qStarts tStarts
      #  0       1     2    3    4      5      6       7      8     9    10    11     12     13    14      15    16     17       18       19      21
-     while (<BL_OUT>) {
+     while (<BLAT>) {
+        last if ($inblat && m/^#.#/);
+        print WBLAT $_ if $wblat;
         chomp;
         my @psl=split(/\t/);
+        die("Error: expect blast alignment for $tseqid but got $psl[13] instead !\n") if $tseqid ne $psl[13];
+        die("Error: unexpected blat strand specification: -+ for $tseqid\n") if ($psl[8] eq '-+');
         my $revmap=($psl[8] eq '+-');
         my ($cds_start, $cds_end); #0-based CDS start, end offsets on the transcript sequence
         my $bscore = (3 * ($psl[0] + $psl[2])) - (3 * $psl[1]) - $psl[4] - $psl[6];
         my $bstatus='gmap';
         my ($bCDstart, $bCDend) = $revmap ? ($psl[14]-$psl[16], $psl[14]-$psl[15]) : ($psl[15], $psl[16]); #0-based PSL start coordinate for the CDS
+        #my ($bCDstart, $bCDend) = ($psl[15], $psl[16]);
         if ($debug) {
            print STDERR ">analyzing $tseqid mapping: $chr:$tstart-$tend|$tstrand CDS:$CDstart-$CDend: blat: $bCDstart-$bCDend\n";
         }
@@ -127,7 +150,12 @@ my $counter=0;
           $cds_end=$bCDend;
           $bstatus='blat';
         }
-        my ($seq, $ofs)=$revmap ? (reverseComplement($tseq), $psl[15]) : ($tseq, $bCDstart);
+        my ($seq, $ofs)=$revmap ? (reverseComplement($tseq), $bCDstart) : ($tseq, $bCDstart);
+        if ($debug) {
+          print STDERR "$psl[8] alignment of qry($psl[11]-$psl[12]) to target($psl[15]-$psl[16])\n";
+          print STDERR "bCDstart=$bCDstart, bCDend=$bCDend\n";
+          print STDERR " seq at offset $ofs:\n".substr($seq, $ofs, 70)."\n";
+        }
         my @cods=unpack('(A3)*', substr($seq, $ofs));
         my $ci=0;
         foreach my $c (@cods) {
@@ -135,11 +163,16 @@ my $counter=0;
           last if exists($stops{$c});
         }
         #stop found at $ofs+$ci*3
-        $ci=scalar(@cods)-1 if $ci==@cods;
+        if ($ci==@cods) {
+           $ci=scalar(@cods)-1 ;
+           print STDERR "\tno stop codon found, \$ci adjusted to $ci\n" if $debug;
+        }
         my $adj_end=$ofs+$ci*3;
-        $adj_end=$psl[14]-$adj_end if $revmap;
+        #$adj_end=$psl[14]-$adj_end if $revmap;
         if ($adj_end!=$cds_end) {
-          #print STDERR "\t$tseqid mapping adjusted ORF end from $cds_end to $adj_end\n";
+          if ($debug) {
+            print STDERR "\t$tseqid mapping adjusted ORF end from $cds_end to $adj_end\n";
+          }
           $cds_end=$adj_end;
           $bscore=3*($cds_end-$cds_start);
           $bstatus.='_adj';
@@ -154,25 +187,45 @@ my $counter=0;
          $tm{$tid}=$td;
        }
        #adjust CDS offsets by original transcript strand:
-       ($cds_start, $cds_end)=($tlen-$cds_end, $tlen-$cds_start) if $tstrand eq '-';
-       push(@{$td}, [$tseqid, $chr, $tstrand, $tstart, $tend, [@ex], $bscore, $cds_start, $cds_end, $bstatus]);
+       my $adjstrand=$tstrand;
+       if ($revmap && $bstatus=~m/^blat/) {
+         $adjstrand = $tstrand eq '-' ? '+' : '-';
+       }
+       ($cds_start, $cds_end)=($tlen-$cds_end, $tlen-$cds_start) if $adjstrand eq '-';
+       push(@{$td}, [$tseqid, $chr, $adjstrand, $tstart, $tend, [@ex], $bscore, $cds_start, $cds_end, $bstatus]);
      }#for each BLAT mapping
     } #reading blat results
-    close(BL_OUT);
-    waitpid($prcid,0);
+    print WBLAT "#.#\n";
+    if (!$inblat) {
+       close(BLAT);
+       waitpid($prcid,0);
+    }
  #-- for now --
  #last if ($counter>5); #for testing now
- } #while FASTA dna records
+ } #while FASTA transcript dna records for the mappings
 }
-unlink($ftmp);
+
+if ($wblat) {
+  close(WBLAT);
+}
+if ($inblat) {
+ close(BLAT);
+} else {
+ unlink($ftmp) unless $debug;
+}
 
 #now pick the best mapping and show the results:
+my $totalcds=0;
 while (my ($tid, $td) = each %tm) {
   my @srtd=sort { $b->[6] <=> $a->[6] } @$td;
   #first entry should be the largest score!
-  my $descr=$tdescr{$tid};
+  my $pd=$tdescr{$tid};
+  die("Error: could not retrieve protein data for $tid\n") unless $pd;
+  my ($plen, $pdescr)=@$pd;
   #cstart and $cend are 0-based, already adjusted for strand
   my ($tseqid, $chr, $strand, $start, $end, $exs, $bscore, $cstart, $cend, $bstatus)=@{$srtd[0]};
+  my $prr=(($cend-$cstart)*100.00)/(3*$plen);
+  $totalcds+=$cend-$cstart;
   my $tacc=0; #accumulating length
   my @cds; #CDS segments to populate
   my ($CDstart, $CDend);
@@ -200,9 +253,13 @@ while (my ($tid, $td) = each %tm) {
  $attrs.=";exonCount=".scalar(@$exs);
  $attrs.=";exons=".join(',', (map { $$_[0].'-'.$$_[1] } @$exs));
  $attrs.=";CDS=$CDstart:$CDend";
- $attrs.=";descr=$descr" if $descr;
+ if ($prr<50.0) {
+    $attrs.= $prr<30.0 ? ';prr=low' : ';prr=poor';
+ }
+ $attrs.=";descr=$pdescr" if $pdescr;
  print join("\t", $chr, 'remap', 'transcript', $start, $end, '.', $strand, '.', $attrs)."\n";
 }
+printf '## Protein Recovery Rate (PRR) : %.2f'."\n", (($totalcds*100.00)/(3*$totalprot));
 
 # --
 if ($outfile) {
@@ -216,4 +273,3 @@ sub reverseComplement {
   $s =~ tr/AaCcTtGgUuMmRrWwSsYyKkVvHhDdBb/TtGgAaCcAaKkYyWwSsRrMmBbDdHhVv/;
   return $s;
  }
-
